requirements.txt
python-dotenv==1.0.1
openai==1.66.5
streamlit==1.44.1
pyvespa==0.34.0
beautifulsoup4==4.12.2
requests==2.31.0
selenium==4.30.0
langchain==0.3.22
fastapi>=0.68.0
uvicorn>=0.15.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5
pydantic>=1.8.2